vllm-API/requirements.txt
# 核心依賴
#vllm
pydantic-settings>=2.0
python-dotenv>=1.0
openai>=1.0
httpx>=0.25

# Web UI 依賴
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# 視覺模型支援（圖片處理）
pillow>=10.0
python-docx
python-pptx

# Benchmark & 工具
rich>=13.0
